[wip] Code review
