+++ astro/package.json
@@ -8,7 +8,7 @@
   "scripts": {
     "//build:ask-ai": "Called explicitly by every script that needs dist/, because Yarn Berry ignores arbitrary pre* hooks. TODO: assumes the Astro CI job invokes one of the build:* scripts rather than astro build directly — confirm when that job is written in DataDog/documentation-ci.",
     "build:ask-ai": "yarn --cwd ../shared/packages/ask-ai build",
-    "//fetch:examples": "Stages the SDK code examples the API docs render. Chained into dev and build the same way build:ask-ai is, since Yarn Berry ignores arbitrary pre* hooks. No-ops when api-code-examples/ is already staged for the current branch; pass --force to re-fetch, or --pins <path> to read a local sdk_versions.json instead of the network. The dev scripts add --best-effort so an unreachable network warns instead of stopping the dev server; no build:* script does, so no build can ship API pages with missing SDK tabs.",
+    "//fetch:examples": "Stages the SDK code examples the API docs render. Chained into dev and build the same way build:ask-ai is, since Yarn Berry ignores arbitrary pre* hooks. No-ops when api-code-examples/ is already staged for the current branch; run `yarn fetch:examples --force` to re-fetch, or pass --pins <path> to read a local sdk_versions.json instead of the network. The dev scripts add --best-effort so an unreachable network warns instead of stopping the dev server; no build:* script does, so no build can ship API pages with missing SDK tabs.",
     "fetch:examples": "node scripts/fetchApiCodeExamples.ts",
     "dev": "yarn build:ask-ai && yarn fetch:examples --best-effort && astro dev --port 4321",
     "dev:proxied": "yarn build:ask-ai && yarn fetch:examples --best-effort && PROXIED=1 astro dev --port 4321",
