+++ astro/scripts/apiCodeExamples/stageHugoLegacyExamples.ts
@@ -1,0 +1,66 @@
+/**
+ * Stages the 148 `.py`/`.rb` files that are *committed* in Hugo rather than
+ * generated by any SDK repo. Five live operations render a Python or Ruby tab
+ * only because of these.
+ *
+ * Enumerated with `git ls-files`, which reads the **index** rather than the
+ * working tree. That is the whole point: it returns the same 148 files whether
+ * or not Hugo has ever been built, and structurally cannot pick up Hugo's build
+ * output. `hugo/.gitignore` ignores the six generated extensions and
+ * deliberately not `.py`/`.rb`, which is why these are tracked at all.
+ */
+
+import path from "node:path";
+import { runGit } from "./git.ts";
+import { HUGO_ROOT } from "./locations.ts";
+import {
+  LEGACY_INDEX_PREFIX,
+  stripHugoApiContentPrefix,
+} from "./pathResolution.ts";
+import { copyIntoStagedTree } from "./stagedTree.ts";
+
+/** Returns the number of legacy files staged. */
+export async function stageHugoLegacyExamples(): Promise<number> {
+  const indexPaths = await listCommittedLegacyExamples();
+  let stagedCount = 0;
+
+  for (const indexPath of indexPaths) {
+    const stagedPath = stripHugoApiContentPrefix(indexPath);
+    if (stagedPath === null) {
+      continue;
+    }
+    await copyIntoStagedTree(path.join(HUGO_ROOT, indexPath), stagedPath);
+    stagedCount += 1;
+  }
+
+  if (stagedCount === 0) {
+    throw new Error(
+      `Found no committed legacy .py/.rb examples in ${HUGO_ROOT}. ` +
+        `Expected ~148; hugo/.gitignore may have started ignoring them.`,
+    );
+  }
+
+  return stagedCount;
+}
+
+async function listCommittedLegacyExamples(): Promise<string[]> {
+  let stdout: string;
+  try {
+    ({ stdout } = await runGit(
+      [
+        "ls-files",
+        "-z",
+        `${LEGACY_INDEX_PREFIX}v*/*/*.py`,
+        `${LEGACY_INDEX_PREFIX}v*/*/*.rb`,
+      ],
+      { cwd: HUGO_ROOT, maxBuffer: 8 * 1024 * 1024 },
+    ));
+  } catch (error) {
+    throw new Error(
+      `Could not list Hugo's committed legacy examples in ${HUGO_ROOT}. ` +
+        `Astro reads them out of Hugo's git index: ${(error as Error).message}`,
+    );
+  }
+
+  return stdout.split("\0").filter((line) => line.length > 0);
+}
